+++ Sources/CoreData/Repository/ObjectListRepository.swift
@@ -20,7 +20,8 @@ public protocol ObjectListRepository {
 
 public extension ObjectListRepository {
 	func refresh(then handler: @escaping (Result<[ObjectType], Error>) -> Void) {
-		handler(.cancelled)
+		assertionFailure("Forgot to implement refresh handler of repository \(Self.self).")
+		handler(.failure(UnimplementedMethod()))
 	}
 }