+++ Sources/CoreData/Repository/Repository.swift
@@ -6,3 +6,6 @@
 /// Put all your repositories in this namespace
 public enum Repository {
 }
+
+/// Internal error thrown by (for example) the `refresh()` functions.
+struct UnimplementedMethod: Error {}